refactor(api): Move autoPlan to config
