+++ lib/actions/form.js
@@ -41,7 +41,7 @@ export function formChanged () {
   return function (dispatch, getState) {
     dispatch(changingForm())
     // TODO: make auto-plan trip a configurable
-    if (queryIsValid(getState().otp) && getState().otp.ui.autoPlan) {
+    if (queryIsValid(getState().otp) && getState().otp.config.autoPlan) {
       dispatch(planTrip())
     }
   }
