+++ lib/index.js
@@ -15,7 +15,7 @@ import ItineraryCarousel from './components/narrative/itinerary-carousel'
 import NarrativeItineraries from './components/narrative/narrative-itineraries'
 import NarrativeItinerary from './components/narrative/narrative-itinerary'
 
-import { setAutoPlan } from './actions/ui'
+import { setAutoPlan } from './actions/config'
 
 import createOtpReducer from './reducers/create-otp-reducer'
