+++ lib/reducers/create-otp-reducer.js
@@ -4,6 +4,10 @@ import update from 'immutability-helper'
 // import { getDefaultQuery } from '../util/state'
 import { getCurrentDate, getCurrentTime } from '../util/time'
 
+const defaultConfig = {
+  autoPlan: true
+}
+
 // TODO: parse query params
 const defaultQuery = {
   type: 'ITINERARY',
@@ -24,13 +28,10 @@ function createOtpReducer (config, initialQuery) {
   const currentQuery = Object.assign(defaultQuery, initialQuery)
   console.log(currentQuery)
   const initialState = {
-    config,
+    config: Object.assign(defaultConfig, config),
     currentQuery,
     searches: [],
-    activeSearch: 0,
-    ui: {
-      autoPlan: true
-    }
+    activeSearch: 0
   }
 
   return (state = initialState, action) => {
@@ -105,7 +106,7 @@ function createOtpReducer (config, initialQuery) {
         return update(state, { activeSearch: { $set: null } })
 
       case 'SET_AUTOPLAN':
-        return update(state, { ui: { autoPlan: { $set: action.payload.autoPlan } } })
+        return update(state, { config: { autoPlan: { $set: action.payload.autoPlan } } })
 
       default:
         return state
